feat: "Implementacion DTO definiendo servicios"

+++ src/main/java/com/crud/crudprueba/repository/ICardRepository.java
@@ -3,6 +3,6 @@ package com.crud.crudprueba.repository;
 import com.crud.crudprueba.entity.Card;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ICardRepository extends CrudRepository<Card, Long>{
+public interface ICardRepository extends CrudRepository<Card, Integer>{
 
 }
